Added a high pass filter for eden resizing
The eden sizing logic that is used when the heap is fully expanded, now has a high pass filter.
This means that predicted improvements for cpu overhead and pause times, must be relatively signficiant (at least 5% improvement) for eden to even bother resizing.

Along with this, a few minor changes have been added, including:
- max heap resize is now `1.25` times the current heap size, rather than `2` times current heap size (which is much too large - the prediction is not accurate)
- Changed the minimum eden % from 1% to 5%. 1% was a bit too low, particularly with slightly different startup logic.
- Each pgc will increase/decrease eden by 5% (up from 3%).
- When heap is not fully expanded, allow pgc pause time to cpu overhead mapping, to suggest negative values for the pause time overhead.
  This is a little bit strange conceptually, but it allows for pause times to be respected significantly better than by preventing overhead from being mapped below 0.
  When the mapped overhead falls below 0, it simply suggests contraction more aggressively.

Performance wise, the biggest change which comes from this commit, is that eden size is significantly more stable when the heap is fully expanded.

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>