Resize heap at GMP endpoints
When a GMP cycle is finished, the subsequent PGC's will free unsused memory which was not marked during GMP.
Once the maximum amount of free memory has been recovered, the heap will be allowed to expand.

This is required, since "tenure" space will slowly be consumed by transient objects, but a heap expansion is not really required in these cases. Expansion should only occur due to real long lived objects

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>