Move MM_HeapSizingData to GlobalVLHGCStats on OMR side
Moving this struct to `GlobalVLHGCStats` is a better spot - It is specific enough to Balanced gc, while being a global structure (as opposed to thread local structure in EnvironmentVLHGC)

This commit also makes use of new heap expansion/contraction reasons. When the heap expands ONLY because eden expanded, the expansion reason now reflects the correct reason. Same logic applies for eden contraction.

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>