Refactoring VGC to print out a line specifically if eden resizes
With this modification, when eden is resizing, a VGC line will be outputted to reflect by how many bytes eden expanded/contract. The "space" value for this, is "eden"

Whether eden contracted/expanded or not, a (posible second) line which will output the net/overall amount by which the heap was resized. This includes both eden, and non-eden resizing, into one single resize line.
The "space" value for this heap resize line is "total heap"

If a user wishes to determine how much non-eden changed, they can simply substract "total heap" resize value, but "eden" resize value.
If non-eden is the only reason for heap expansion, then "non-eden" resize amount is the same as "total heap"

In a perfect world, the entire <heap-resize> stanza, would be a nested/embedded style, and within that nesting, would include values for eden, and non-eden. This will be the subject of another issue on github.

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>